Download test roms when running tests
Test roms were previously part of the repository as a submodule. This
made the repository unecessarily big for those who just want to install
the emulator by cloning the repository. Even worse, the repository was
so big that go install refused to download it...

Instead, we now lazy-download the test roms to ./tests/nes-test-roms the
first time the user runs go test.