CPU: simplify Running a single frame
This is in preparation for the various features of the "game panel".
The user will be able to perform "Restart", "Pause" or "Reset" on a
running at any moment, but in fact we'll just handle this in between 2
frames since we don't want to introduce concurrent code in the CPU/hw
execution.