ui: saving and loading of battery packed ram
When the emulator stops, any save ram gets copied from the emulating
session temporary directory to the $CONFIG/saveram directory.

When starting a rom, the expected battery packed ram save file path is
looked upon in the $CONFIG, if found, that buffer is used in lieu of
the starting zero'ed buffer.