Mapper snapshots save/restore (#65)
* all: add package doc comments

* hw: remove warning in generated cpu code

* vendor: bump msgp to v1.3.0

* emu+hw: speedup NES snapshot creation

When creating/restoring snapshot, RAM has just to be copied,
as-is, so we use a pointer to array conversion in order to
reduce allocation and get a 10% speedup on NES.SaveSnapshot
micro benchmark.

* hw/mappers: remove useless MapBank in base.init()

* go.mod: bump to go1.25

* emu: fix bug in LoadSnapshot

Since hwio.Mem uses an uintptr, we can just change the CPU RAM slice.
We need to copy the data. Or else we'd need to update the uintptr in
hwio.Mem.

* Disable CPU tracing during run-ahead frames

* hw/snapshot: move Input and DMA state into CPU

* hw/mappers: save/restore mapper state for frame run-ahead

* Cosmetics

* hw: add dummy cycle to DMA state

* hw: add hwio.Reg8 values to snapshot state

* hw/mappers: simplify state serialization

* hw/mappers: reorganize files

* hw/mappers: simplify SetState